backend/requirements.txt
fastapi
uvicorn
crewai
yfinance
pandas
ta
python-dotenv
pydantic
pydantic-settings
requests
scikit-learn
supabase

# LLM providers (auto-switching — add whichever you have a key for)
langchain-openai
langchain-google-genai
langchain-groq
langchain-huggingface

# Agent Chat & RAG
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
crewai-tools
